pyproject.toml
[tool.black]
line-length = 160

[tool.poetry]
name = "it2zbot"
version = "0.1.0"
description = ""
authors = ["Eulentier161"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
pyaml = "^21.10.1"
python-dotenv = "^0.20.0"
uvloop = {version = "^0.16.0", markers = 'sys_platform != "win32"'}
SQLAlchemy = "^1.4.34"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
